Implemented correct deltaTime
Deltatime is now send to each update method in seconds as a double.

+++ Sandbox/Engine/Module System/Module.cs
@@ -37,7 +37,7 @@ namespace ZEngine.Core
 
         }
 
-        public virtual void Update(float deltaTime)
+        public virtual void Update(double deltaTime)
         {
 
         }
