+++ Zayker Game Engine/Zayker Game Engine/Core/Game.cs
@@ -12,7 +12,7 @@ public static class Game
 
     }
 
-    public static void Update(float deltaTime)
+    public static void Update(double deltaTime)
     {
 
     }
