+++ Zayker Game Engine/Zayker Game Engine/Modules/example/ExampleModule.cs
@@ -22,7 +22,7 @@ namespace ZEngine.ExampleModule
             base.OnDisable();
         }
 
-        public override void Update(float deltaTime)
+        public override void Update(double deltaTime)
         {
             base.Update(deltaTime);
         }
